Fixes/fix robotic arm model transforms (#53)
* feat: added two new robot arm models "panda" and "iiwa14"

TODO:
- Add last one "ur5e"
- Add default pos (loaded pos) to json

* feat: added ur5e robot model and specific DSL scripts for testing and evaluation

* fixes: fixed Panda robot model and specific DSL scripts for testing and evaluation, along  with specific damping and friction values

* feat: added iiwa robot model and specific DSL scripts for testing and evaluation

TODO:
- Figure out why iiwa14 model doesnt rotate correctly, and fix existing kinematic updaters to visually work with different model types.
- ADD MATERIALS
- Already started, but beging report.

* feat: added similar scripts for Z1 robotic arm model, more realisitc comparisons with the same integrators!

TODO:
- Add a GENERAL script, acting as a constant.

* feat: extra scripts

* Remove: old scripts removed from repo

* fixes: iiwa14 fixes, added intertial info

TODO:
- Add colours!!!!!!

* fixes: Updated robotic loading and visual kineamtics with robust logic for different robot joint types, and general safety measures

* fixes

* fixes

* feat: add materials per robotic arm links (Colour only for now), move non-visual var types from float to double

TODO:
-Update calls of said methods using floats for what now is double type params.

* fixes: robot visual fk for mixed mesh frame convention

NOTES (IMPORTANT):
- Some robot assets (e.g. Z1 CAD-derived meshes) encode their link frames directly in mesh local space, while others (URDF-style, e.g. Panda) expect FK to be applied at the object level
- Engine-aligned models: bake FK directly into mesh local transform
- URDF-style models: apply FK + visual origin to object transform
- MAY REVISE LATER

* feat: add UR5e integrator stress test script, along with basic colours for each arm

+++ Sim_Engine/EngineLib/include/Analysis/Telemetry.h
@@ -20,25 +20,25 @@ namespace diagnostics {
 	// Struct for joint telemetry data
 	struct ENGINE_API JointTelemetry{
 		// Actual data
-		float thetaRad = 0.0f;		 // Joint angle in radians
-		float omegaRad_s = 0.0f;	 // Joint angular velocity in radians per second
-		float eta = 0.0f;			 // 
+		double thetaRad = 0.0f;		 // Joint angle in radians
+		double omegaRad_s = 0.0f;	 // Joint angular velocity in radians per second
+		double eta = 0.0f;			 // 
 
 		// Additional dynamics data
-		float torqueNm = 0.0f;		 // Joint torque (N·m)
-		float damping = 0.0f;		 // Joint damping coefficient  (kg·m²/s)
-		float friction = 0.0f;		 // Joint friction coefficient (Coulomb friction - N·m)
-		float effort = 0.0f;		 // Normalized effort (0 to 1)
+		double torqueNm = 0.0f;		 // Joint torque (N·m)
+		double damping = 0.0f;		 // Joint damping coefficient  (kg·m²/s)
+		double friction = 0.0f;		 // Joint friction coefficient (Coulomb friction - N·m)
+		double effort = 0.0f;		 // Normalized effort (0 to 1)
 
 		// Reference data
-		float thetaRefRad = 0.0f;	 // Reference joint angle in radians
-		float omegaRefRad_s = 0.0f;	 // Reference joint angular velocity in radians per second
-		float alphaRefRad_s2 = 0.0f; // Reference joint angular acceleration in radians per second squared
+		double thetaRefRad = 0.0f;	 // Reference joint angle in radians
+		double omegaRefRad_s = 0.0f;	 // Reference joint angular velocity in radians per second
+		double alphaRefRad_s2 = 0.0f; // Reference joint angular acceleration in radians per second squared
 
 		// Trajectory data
-		float traj_q = 0.0f;		 // Trajectory joint position
-		float traj_qd = 0.0f;		 // Trajectory joint velocity
-		float traj_qdd = 0.0f;		 // Trajectory joint acceleration
+		double traj_q = 0.0f;		 // Trajectory joint position
+		double traj_qd = 0.0f;		 // Trajectory joint velocity
+		double traj_qdd = 0.0f;		 // Trajectory joint acceleration
 		bool traj_active = false;	 // Trajectory active state for a joint
 
 		// Limit clamping flags
@@ -53,8 +53,8 @@ namespace diagnostics {
 		std::vector<JointTelemetry> j;	// Vector of joint telemetry data
 
 		// Summary statistics (precomputed to relieve analysis load)
-		float err_rms = 0.0f;			// RMS error across all joints
-		float err_max = 0.0f;			// Maximum error across all joints
+		double err_rms = 0.0f;			// RMS error across all joints
+		double err_max = 0.0f;			// Maximum error across all joints
 		int clamp_sum = 0;				// Sum of clamping events across all joints
 		int worst_joint = -1;			// Index of the joint with max |e|
 	};
